src/backend/requirements.txt
fastapi
uvicorn
unsloth
transformers
accelerate
datasets
trl
Pillow
psutil
requests
google-generativeai
openai
transformers
torch
huggingface_hub
pydantic
python-multipart
scikit-learn
sqlite3
sentence-transformers
bert-score
